move plugin to org.logstashplugins package
