+++ build.gradle
@@ -7,7 +7,7 @@ apply from: LOGSTASH_CORE_PATH + "/../rubyUtils.gradle"
 // ===========================================================================
 // plugin info
 // ===========================================================================
-group                      'org.logstash.javaapi' // must match the package of the main plugin class
+group                      'org.logstashplugins' // must match the package of the main plugin class
 version                    "${file("VERSION").text.trim()}" // read from required VERSION file
 description                = "Example Java output implementation"
 pluginInfo.licenses        = ['Apache-2.0'] // list of SPDX license IDs
